Fix severe performance issue with ::getTFPrefix()
Only call getTFPrefix() once in constructor and cache result, instead of
calling it every time we publish TF updates. Because it searches through
the rosparam list which involves lots of comms with roscore, this was
causing TF updates to always be delayed a few millisecs and sometimes
several hundred millisecs, even if this node is run with realtime
priority! This fixes the problem.